added search code
The files necessary to a functioning search bar. Note: home.html needs to be in a nested folder that looks like core/templates/core/home.html. urls.py and views.py are updated with search_view() and the path on the urls. Both are files that belong to core as well, so they only need to be updated there